requirements.txt
PyYAML==6.0
langchain
openai
kor
fastapi
uvicorn[standard]
pydantic